Split train & test files
