+++ models/train_power.py
@@ -3,7 +3,6 @@ import copy
 import torch
 from torch import nn
 import torch.distributed as dist
-from datetime import datetime
 import warnings
 from era5_data import utils, utils_data
 from era5_data.config import cfg
@@ -335,42 +334,3 @@ def validate(
             epochs_since_last_improvement += 1
 
     return val_loss, best_model, epochs_since_last_improvement
-
-
-def test(test_loader, model, device, res_path):
-    aux_constants = load_constants(device)
-    for id, data in enumerate(test_loader, 0):
-        timestamp = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-        print(f"[{timestamp}] predict on {id}")
-        (
-            input_test,
-            input_surface_test,
-            target_power_test,
-            target_upper_test,
-            target_surface_test,
-            periods_test,
-        ) = data
-
-        input_test, input_surface_test, target_power_test = (
-            input_test.to(device),
-            input_surface_test.to(device),
-            target_power_test.to(device),
-        )
-        model.eval()
-        output_power_test, output_surface_test = model_inference(
-            model, input_test, input_surface_test, aux_constants
-        )
-        lsm_expanded = load_land_sea_mask(output_power_test.device)
-        output_power_test = output_power_test * lsm_expanded
-        target_time = periods_test[1][0]
-        png_path = os.path.join(res_path, "png")
-        utils.mkdirs(png_path)
-        visualize(
-            output_power_test,
-            target_power_test,
-            input_surface_test,
-            output_surface_test,
-            target_surface_test,
-            target_time,
-            png_path,
-        )
